MySQL: Store float (xml schema) as double in mysql
git-svn-id: file:///www/metaways/svn.tine20.org/repo/trunk@9911 8d11cdf3-9d2f-0410-bbb5-e098f187f06b

+++ tine20/Setup/Backend/Mysql.php
@@ -401,7 +401,7 @@ class Setup_Backend_Mysql extends Setup_Backend_Abstract
                 break;
             
             case 'float':
-                $buffer[] = 'float';
+                $buffer[] = 'double';
                 break;
             
             case 'decimal':
